requirements.txt
torch>=1.11.0
torchvision>=0.12.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
scipy>=1.7.0
imbalanced-learn>=0.8.0  # 用于SMOTE
matplotlib>=3.3.0
seaborn>=0.11.0
tensorboard>=2.8.0
tqdm>=4.60.0
pyyaml>=5.4.0
jupyter>=1.0.0